Move some deps to devDependencies
I just pulled this through npm and it ended up pulling coffee script and should as well.  I'm assuming they are not runtime dependencies.

I *think* uglify-js is also a build-time concern, so devDependencies should be sufficient?  

I'm not familiar enough with requirejs or how you are using line-reader or strscan to know where those belong, but I'm pretty sure that these 3 don't need to be pulled down into my project that is just trying to use walltime.